You call it if/else, I call it machine learning.

+++ include/sentiment.hpp
@@ -9,9 +9,9 @@ namespace sentiment
 {
 bool is_negative(const std::string&);
 
-int value(const std::string&);
+float value(const std::string&);
 
-int default_negative_threshold(const size_t len);
+float default_negative_threshold(const size_t len);
 
 std::string key(const std::string& in);
 }
